+++ languages/translations/en.json
@@ -369,7 +369,7 @@
       "pending": "Pending",
       "confirmed": "確認済み",
       "delivering": "配信中",
-      "completes": "完了",
+      "completed": "完了",
       "canceled": "キャンセル",
       "date_delivery": "配達日",
       "date_delivered": "配達日"
